LayoutTests/imported/w3c/web-platform-tests/mathml/presentation-markup/operators/operator-dictionary-combining-default.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Operator dictionary (combining char, default category)</title>
<link rel="stylesheet" type="text/css" href="/fonts/ahem.css" />
<link rel="help" href="https://w3c.github.io/mathml-core/#operator-fence-separator-or-accent-mo">
<link rel="help" href="https://w3c.github.io/mathml-core/#dictionary-based-attributes">
<link rel="help" href="https://w3c.github.io/mathml-core/#operator-dictionary">
<link rel="help" href="https://w3c.github.io/mathml-core/#layout-of-mrow">
<link rel="stylesheet" href="./support/operator-dictionary-tests.css"/>
<meta name="assert" content="Operators with unsupported combining characters or multiple characters fall back to Default category">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/mathml/support/feature-detection.js"></script>
<style>
  mo {
      color: blue;
  }
  mn {
      background: black;
  }
</style>
<script>
  setup({ explicit_done: true });
  window.addEventListener("load", runTests);

  function spaceBetween(element, i, j) {
      return element.children[j].getBoundingClientRect().left -
          element.children[i].getBoundingClientRect().right;
  }
  function runTests() {
      var epsilon = 1;

      // Operators with unsupported combining characters or multiple combining
      // characters should fall back to Default category (0em spacing).
      Array.from(document.getElementsByClassName("default")).forEach(element => {
          test(function() {
              assert_true(MathMLFeatureDetection.has_operator_spacing());
              var lspace = spaceBetween(element, 0, 1);
              var rspace = spaceBetween(element, 1, 2);
              assert_approx_equals(lspace, 0, epsilon, "lspace");
              assert_approx_equals(rspace, 0, epsilon, "rspace");
          }, `Default spacing for ${element.getAttribute("data-description")}`);
      });

      done();
  }
</script>
</head>
<body>
  <div id="log"></div>
  <!-- U+0303 COMBINING TILDE is not a recognized combining overlay -->
  <p>
    <math class="default" data-description="= + unsupported combining char U+0303">
      <mn>&nbsp;</mn>
      <mo>=&#x303;</mo>
      <mn>&nbsp;</mn>
    </math>
  </p>
  <!-- Three characters: base + two combining overlays -->
  <p>
    <math class="default" data-description="= + U+0338 + U+20D2 (too many characters)">
      <mn>&nbsp;</mn>
      <mo>=&#x338;&#x20D2;</mo>
      <mn>&nbsp;</mn>
    </math>
  </p>
  <!-- Two regular (non-combining) characters -->
  <p>
    <math class="default" data-description="++ (two non-combining characters)">
      <mn>&nbsp;</mn>
      <mo>++</mo>
      <mn>&nbsp;</mn>
    </math>
  </p>
</body>
</html>
